chore: rename audio factory to tensor factory
