+++ src/spira_training/shared/adapters/pytorch/model_trainer/implementations/basic_pytorch_audio_factory.py
@@ -3,7 +3,7 @@ from src.spira_training.shared.adapters.pytorch.models.pytorch_tensor import (
 )
 from src.spira_training.shared.core.models.audio import Audio
 import torch
-from src.spira_training.shared.adapters.pytorch.model_trainer.interfaces.pytorch_audio_factory import (
+from src.spira_training.shared.adapters.pytorch.model_trainer.interfaces.pytorch_tensor_factory import (
     PytorchTensorFactory,
 )
