+++ src/spira_training/shared/adapters/pytorch/model_trainer/implementations/pytorch_dataloader_factory.py
@@ -4,7 +4,7 @@ import torch
 
 from src.spira_training.shared.adapters.pytorch.models.pytorch_label import PytorchLabel
 
-from src.spira_training.shared.adapters.pytorch.model_trainer.interfaces.pytorch_audio_factory import (
+from src.spira_training.shared.adapters.pytorch.model_trainer.interfaces.pytorch_tensor_factory import (
     PytorchTensorFactory,
 )
