+++ src/spira_training/shared/adapters/pytorch/model_trainer/interfaces/pytorch_tensor_factory.py
