+++ src/spira_training/shared/core/audio_processor_factory.py
@@ -1,4 +1,4 @@
-from src.spira_training.shared.adapters.pytorch.model_trainer.interfaces.pytorch_audio_factory import (
+from src.spira_training.shared.adapters.pytorch.model_trainer.interfaces.pytorch_tensor_factory import (
     PytorchTensorFactory,
 )
 from src.spira_training.apps.feature_engineering.configs.audio_processor_config import (
